Move silence_logging into logging_util.py
